requirements.txt
pandas
pyyaml
tabulate
python-dotenv
openai
requests
